squash: move into participant actions
