+++ modules/UI/util/MessageHandler.js
@@ -1,6 +1,4 @@
-/* global $, APP, interfaceConfig */
-import throttle from 'lodash/throttle';
-
+/* global $, APP */
 const logger = require('jitsi-meet-logger').getLogger(__filename);
 
 import jitsiLocalStorage from '../../util/JitsiLocalStorage';
@@ -114,74 +112,6 @@ function dontShowAgainSubmitFunctionWrapper(options, submitFunction) {
 function isDontShowAgainEnabled(options) {
     return typeof options === 'object';
 }
-
-/**
- * By default how long a notification should display before being automatically
- * dismissed.
- *
- * @private
- * @type {number}
- */
-const NOTIFICATION_DISMISS_TIMEOUT = 2500;
-
-/**
- * An array of names of participants that have joined the conference. The array
- * is replaced with an empty array as notifications are displayed.
- *
- * @private
- * @type {string[]}
- */
-let joinedParticipantsNames = [];
-
-/**
- * A throttled internal function that takes the internal list of participant
- * names, {@code joinedParticipantsNames}, and triggers the display of a
- * notification informing of their joining.
- *
- * @private
- * @type {Function}
- */
-const throttledNotifyParticipantConnected = throttle(() => {
-    const joinedParticipantsCount = joinedParticipantsNames.length;
-
-    let notificationProps;
-
-    if (joinedParticipantsCount >= 3) {
-        notificationProps = {
-            titleArguments: {
-                name: joinedParticipantsNames[0],
-                count: joinedParticipantsCount - 1
-            },
-            titleKey: 'notify.connectedThreePlusMembers'
-        };
-    } else if (joinedParticipantsCount === 2) {
-        notificationProps = {
-            titleArguments: {
-                first: joinedParticipantsNames[0],
-                second: joinedParticipantsNames[1]
-            },
-            titleKey: 'notify.connectedTwoMembers'
-        };
-    } else if (joinedParticipantsCount) {
-        notificationProps = {
-            titleArguments: {
-                name: joinedParticipantsNames[0]
-            },
-            titleKey: 'notify.connectedOneMember'
-        };
-    }
-
-    if (notificationProps) {
-        APP.store.dispatch(
-            showNotification(
-                Notification,
-                notificationProps,
-                NOTIFICATION_DISMISS_TIMEOUT));
-    }
-
-    joinedParticipantsNames = [];
-
-}, 500, { leading: false });
 
 const messageHandler = {
     OK: 'dialog.OK',
@@ -562,7 +492,7 @@ const messageHandler = {
             cls,
             messageKey,
             messageArguments,
-            timeout = NOTIFICATION_DISMISS_TIMEOUT) {
+            timeout = 2500) {
         APP.store.dispatch(
             showNotification(
                 Notification,
@@ -573,21 +503,6 @@ const messageHandler = {
                     title: displayName
                 },
                 timeout));
-    },
-
-    /**
-     * Queues the display of a notification of a participant having connected to
-     * the meeting. The notifications are batched so that quick consecutive
-     * connection events are shown in one notification.
-     *
-     * @param {string} displayName - The name of the participant that connected.
-     * @returns {void}
-     */
-    notifyParticipantConnected(displayName) {
-        joinedParticipantsNames.push(
-            displayName || interfaceConfig.DEFAULT_REMOTE_DISPLAY_NAME);
-
-        throttledNotifyParticipantConnected();
     },
 
     /**
